update bevy version

+++ src/board.rs
@@ -58,7 +58,7 @@ fn enter_setup(mut query: Query<(&Transform, &mut Phases<BoardMode>), With<Tile>
     for (transform, mut phases) in &mut query {
         let offset = transform.translation.x + transform.translation.z;
         phases.reset(vec![
-            Phase::new(BoardMode::Waiting, 0.4 + offset as f32 * 0.05),
+            Phase::new(BoardMode::Waiting, 0.4 + offset * 0.05),
             Phase::new(BoardMode::Entering, 1.0),
             Phase::new(BoardMode::Shown, 0.0), // final
         ]);
@@ -69,7 +69,7 @@ fn enter_teardown(mut query: Query<(&Transform, &mut Phases<BoardMode>), With<Ti
     for (transform, mut phases) in &mut query {
         let offset = transform.translation.x + transform.translation.z;
         phases.reset(vec![
-            Phase::new(BoardMode::Waiting, 0.1 + offset as f32 * 0.05),
+            Phase::new(BoardMode::Waiting, 0.1 + offset * 0.05),
             Phase::new(BoardMode::Exiting, 1.0),
             Phase::new(BoardMode::Hidden, 0.0), // final
         ]);
